% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_add_dataset}
\alias{sip_add_dataset}
\title{SIP Metastore Add Dataset}
\usage{
sip_add_dataset(output_path, output_format, output_schema, output_rows = 0,
  component = "RComponent", script, desc = "", created_by = NULL,
  project_id, batch_id = "", input_paths = "", input_formats = "",
  input_ids = "", input_names = "", started = NULL, finished = NULL,
  status = "SUCCESS", hostname, token)
}
\arguments{
\item{output_path}{output dataset file path}

\item{output_format}{output dataset format. accepts parquet, JSON, csv and
rds}

\item{output_schema}{output dataset sechema. Should be a list with name and
type elements. list should not be named}

\item{component}{component name used to create dataset. Default is RComponent}

\item{script}{path to executable R script}

\item{desc}{description of dataset. Default is empty string}

\item{created_by}{user name. Default is system user name}

\item{project_id}{project id string}

\item{batch_id}{batch id string. Default is empty string}

\item{input_paths}{vector of input datasets paths. Default is empty string}

\item{input_ids}{vector of input dataset ids. Default is empty string}

\item{input_names}{vectore of input dataset names. Default is empty string}

\item{started}{string with starting time. Format should be YYYYMMDD-HHMMSS.
Default is NULL which inserts current time}

\item{finished}{string with starting time. Format should be YYYYMMDD-HHMMSS}

\item{status}{component status. default is 'SUCCESS'. Default is NULL which
inserts current time}

\item{hostname}{hostname only of saw security location. No pathing required}

\item{token}{valid SIP token. Result of sip_authenticate function}

\item{row_count}{ouput dataset record count. Default is NULL}
}
\value{
POST response content
}
\description{
Adds a dataset record to the SIP Metastore
}
\details{
POST request to
/saw/services/internal/workbench/projects/{project_id}/datasets/create
}
