% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_authenticate}
\alias{sip_authenticate}
\title{SIP Metastore Authenticate Function}
\usage{
sip_authenticate(login, password, hostname)
}
\arguments{
\item{login}{master login character}

\item{password}{password character}

\item{hostname}{hostname only of saw security location. No pathing required
Ex: hostname = "https://saw-bda-cert-vaste.sncrcorp.net"}
}
\value{
aToken authentication string
}
\description{
Generates authentication token for metastore communication
}
\details{
POST request to saw/security/doAuthenticate
}
