% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_get_datasets}
\alias{sip_get_datasets}
\title{SIP Metastore Get Datasets Function}
\usage{
sip_get_datasets(project_id, hostname, token)
}
\arguments{
\item{project_id}{id of project to retrieve datasets from}

\item{hostname}{hostname only of saw security location. No pathing required
Ex: hostname = "https://saw-bda-cert-vaste.sncrcorp.net"}

\item{token}{valid SIP token. Result of sip_authenticate function}
}
\value{
list with datasets metadata
}
\description{
Returns all datasets in Workbench Metastore for a given project
}
\details{
Get request to saw/services/internal/workbench/projects/{project_id}/datasets
}
