% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-install.R
\name{sip_package_install}
\alias{sip_package_install}
\title{Install R Package Function}
\usage{
sip_package_install(packages, lib_path = NULL, dependecies = TRUE,
  repo = "https://cloud.r-project.org")
}
\arguments{
\item{packages}{list of R packges. name of element should be package name.
version is element value}

\item{lib_path}{optional input for library path}

\item{dependecies}{option to install package dependecies}

\item{repo}{repo url to download from}
}
\description{
Installs list of R packages. Checks for pre-installation and correct version
}
